refactor: split UI styles and export analysis helpers
Group feature-style extraction, export format/filter cleanup, and CLI version-analysis decomposition.
