+++ apps/desktop/src/views/mcp/McpManagerView.vue
@@ -6,6 +6,7 @@ import McpAddServerModal from "@/components/mcp/McpAddServerModal.vue";
 import McpServerCard from "@/components/mcp/McpServerCard.vue";
 import McpTokenSummary from "@/components/mcp/McpTokenSummary.vue";
 import { browseForFile } from "@/composables/useBrowseDirectory";
+import "@/styles/features/mcp-manager.css";
 import { useMcpStore } from "@/stores/mcp";
 
 const store = useMcpStore();
@@ -268,355 +269,3 @@ async function handleImport() {
   </div>
   </PageShell>
 </template>
-
-<style scoped>
-.mcp-manager-view {
-  display: flex;
-  flex-direction: column;
-  gap: 0;
-}
-
-/* ─── Buttons ─── */
-.btn-add-server {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 7px 14px;
-  background: var(--gradient-accent);
-  color: white;
-  border: none;
-  border-radius: var(--radius-md);
-  font-size: 0.8125rem;
-  font-weight: 600;
-  cursor: pointer;
-  box-shadow: 0 1px 6px rgba(99, 102, 241, 0.35);
-  transition: all var(--transition-fast);
-  letter-spacing: -0.01em;
-}
-
-.btn-add-server:hover {
-  box-shadow: 0 3px 14px rgba(99, 102, 241, 0.45);
-  transform: translateY(-1px);
-}
-
-.btn-add-server:active {
-  transform: translateY(0);
-}
-
-.btn-add-server svg {
-  width: 14px;
-  height: 14px;
-}
-
-.btn-import {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 7px 14px;
-  background: var(--canvas-subtle);
-  color: var(--text-secondary);
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  font-size: 0.8125rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all var(--transition-fast);
-}
-
-.btn-import:hover {
-  color: var(--text-primary);
-  border-color: var(--border-accent);
-  background: var(--neutral-subtle);
-}
-
-/* ─── Stats Strip ─── */
-.stats-strip {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  padding: 10px 0 2px;
-  flex-wrap: wrap;
-}
-
-.stat-chip {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  font-size: 0.75rem;
-  font-weight: 500;
-  color: var(--text-secondary);
-  letter-spacing: 0.01em;
-}
-
-.stat-dot {
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-
-.stat-dot.installed { background: var(--accent-fg); }
-.stat-dot.active    { background: var(--success-fg); }
-.stat-dot.paused    { background: var(--warning-fg); }
-.stat-dot.error     { background: var(--danger-fg); }
-
-.stat-sep {
-  color: var(--text-placeholder);
-  font-size: 0.625rem;
-  user-select: none;
-}
-
-/* ─── Filter Bar ─── */
-.filter-bar {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 16px 0 20px;
-  flex-wrap: wrap;
-}
-
-.search-container {
-  position: relative;
-  display: flex;
-  align-items: center;
-  max-width: 260px;
-  flex: 1;
-}
-
-.search-icon {
-  position: absolute;
-  left: 10px;
-  color: var(--text-placeholder);
-  pointer-events: none;
-}
-
-.search-input {
-  width: 100%;
-  padding: 7px 10px 7px 32px;
-  background: var(--canvas-subtle);
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  color: var(--text-primary);
-  font-size: 0.8125rem;
-  outline: none;
-  transition: border-color var(--transition-fast);
-}
-
-.search-input:focus {
-  border-color: var(--accent-emphasis);
-}
-
-.search-input::placeholder {
-  color: var(--text-placeholder);
-}
-
-.btn-icon-sm {
-  width: 32px;
-  height: 32px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: var(--canvas-subtle);
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  color: var(--text-secondary);
-  cursor: pointer;
-  transition: color var(--transition-fast), border-color var(--transition-fast);
-}
-
-.btn-icon-sm:hover {
-  color: var(--text-primary);
-  border-color: var(--accent-emphasis);
-}
-
-.btn-icon-sm:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.spinning {
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
-}
-
-/* ─── Section Heading ─── */
-.section-heading {
-  font-size: 0.6875rem;
-  font-weight: 600;
-  color: var(--text-tertiary);
-  text-transform: uppercase;
-  letter-spacing: 0.06em;
-  margin-bottom: 12px;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.section-heading::after {
-  content: "";
-  flex: 1;
-  height: 1px;
-  background: var(--border-default);
-}
-
-/* ─── Tag Filters ─── */
-.tag-filters {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 6px;
-  padding-bottom: 8px;
-}
-
-.tag-filter-btn {
-  padding: 3px 10px;
-  font-size: 0.75rem;
-  border-radius: var(--radius-full);
-  background: var(--neutral-subtle);
-  color: var(--text-tertiary);
-  border: 1px solid transparent;
-  cursor: pointer;
-  transition:
-    background-color var(--transition-fast),
-    color var(--transition-fast),
-    border-color var(--transition-fast);
-}
-
-.tag-filter-btn:hover {
-  color: var(--text-secondary);
-  background: var(--neutral-muted);
-}
-
-.tag-filter-btn.active {
-  background: var(--accent-subtle);
-  color: var(--accent-fg);
-  border-color: var(--accent-emphasis);
-}
-
-/* ─── Error ─── */
-.error-banner {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-  padding: 10px 14px;
-  background: var(--danger-subtle);
-  border: 1px solid var(--danger-muted);
-  border-radius: var(--radius-md);
-  color: var(--danger-fg);
-  font-size: 0.8125rem;
-}
-
-.error-dismiss {
-  background: none;
-  border: none;
-  color: var(--danger-fg);
-  cursor: pointer;
-  font-size: 1.125rem;
-  padding: 0 4px;
-}
-
-/* ─── Loading ─── */
-.loading-state {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  padding: 48px 0;
-  color: var(--text-tertiary);
-  font-size: 0.875rem;
-}
-
-.loading-spinner {
-  width: 20px;
-  height: 20px;
-  border: 2px solid var(--border-default);
-  border-top-color: var(--accent-emphasis);
-  border-radius: 50%;
-  animation: spin 0.8s linear infinite;
-}
-
-/* ─── Empty State ─── */
-.empty-state {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 60px 24px;
-}
-
-.empty-hero {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 16px;
-  width: 100%;
-  text-align: center;
-}
-
-.empty-icon {
-  color: var(--text-tertiary);
-  opacity: 0.4;
-  margin-bottom: 4px;
-}
-
-.empty-title {
-  font-size: 1.25rem;
-  font-weight: 700;
-  color: var(--text-primary);
-  margin: 0;
-  letter-spacing: -0.02em;
-}
-
-.empty-description {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-  margin: 0;
-  line-height: 1.6;
-  max-width: 480px;
-}
-
-.empty-actions {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  margin-top: 8px;
-}
-
-.empty-features {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  margin-top: 20px;
-  padding-top: 20px;
-  border-top: 1px solid var(--border-muted);
-  width: 100%;
-}
-
-.feature-item {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 0.8125rem;
-  color: var(--text-tertiary);
-}
-
-.feature-dot {
-  width: 5px;
-  height: 5px;
-  border-radius: 50%;
-  background: var(--accent-fg);
-  flex-shrink: 0;
-  opacity: 0.6;
-}
-
-/* ─── Server Grid ─── */
-.server-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(310px, 1fr));
-  gap: 12px;
-  margin-bottom: 36px;
-}
-</style>
