+++ apps/desktop/src/views/skills/SkillsManagerView.vue
@@ -9,6 +9,7 @@ import { Brain } from "lucide-vue-next";
 import { computed, onMounted, ref } from "vue";
 import SkillCard from "@/components/skills/SkillCard.vue";
 import SkillImportWizard from "@/components/skills/SkillImportWizard.vue";
+import "@/styles/features/skills-manager.css";
 import { useSkillsStore } from "@/stores/skills";
 
 const store = useSkillsStore();
@@ -71,6 +72,7 @@ async function handleDeleteSkill(dir: string) {
 
 <template>
   <PageShell>
+    <div class="skills-manager-view">
     <PageHeader title="Skills">
         <template #icon>
           <svg viewBox="0 0 16 16" fill="none" stroke="currentColor" stroke-width="1.5" width="16" height="16">
@@ -241,441 +243,6 @@ async function handleDeleteSkill(dir: string) {
         @close="showImportWizard = false"
         @imported="handleImported"
       />
+    </div>
   </PageShell>
 </template>
-
-<style scoped>
-/* ── Stats Strip ─────────────────────────────────────────── */
-.stats-strip {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  padding: 10px 0 2px;
-  flex-wrap: wrap;
-}
-
-.stat-chip {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  font-size: 0.75rem;
-  font-weight: 500;
-  color: var(--text-secondary);
-}
-
-.stat-dot {
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-
-.stat-dot--installed {
-  background: var(--accent-fg);
-}
-
-.stat-dot--global {
-  background: var(--accent-fg);
-}
-
-.stat-dot--project {
-  background: var(--success-fg);
-}
-
-.stat-dot--active {
-  background: var(--success-fg);
-}
-
-.stat-sep {
-  color: var(--text-tertiary);
-  font-size: 0.625rem;
-  user-select: none;
-}
-
-/* ── Token Info ──────────────────────────────────────────── */
-.token-info {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 8px 14px;
-  margin-top: 8px;
-  background: var(--accent-subtle, rgba(99, 102, 241, 0.04));
-  border: 1px solid rgba(99, 102, 241, 0.12);
-  border-radius: var(--radius-md);
-  font-size: 0.75rem;
-  color: var(--text-secondary);
-}
-
-.token-info svg {
-  width: 14px;
-  height: 14px;
-  color: var(--accent-fg);
-  flex-shrink: 0;
-  opacity: 0.7;
-}
-
-.token-info__text {
-  flex: 1;
-}
-
-.token-info code {
-  font-family: "JetBrains Mono", ui-monospace, monospace;
-  font-size: 0.6875rem;
-  color: var(--accent-fg);
-  background: var(--accent-muted);
-  padding: 1px 5px;
-  border-radius: 3px;
-}
-
-.token-info__bar {
-  width: 80px;
-  height: 6px;
-  background: var(--canvas-inset);
-  border-radius: 3px;
-  overflow: hidden;
-  flex-shrink: 0;
-}
-
-.token-info__bar-fill {
-  height: 100%;
-  border-radius: 3px;
-  background: var(--accent-emphasis);
-  transition: width 0.3s ease;
-}
-
-/* ── Filter Row ──────────────────────────────────────────── */
-.filter-row {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  margin: 16px 0 20px;
-  flex-wrap: wrap;
-}
-
-/* ── Scope Segmented Control ─────────────────────────────── */
-.scope-segmented {
-  display: inline-flex;
-  gap: 0;
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  overflow: hidden;
-  flex-shrink: 0;
-}
-
-.scope-seg-btn {
-  padding: 6px 16px;
-  height: 32px;
-  font-size: 0.8125rem;
-  font-weight: 500;
-  color: var(--text-tertiary);
-  background: var(--canvas-default, var(--canvas-subtle));
-  border: none;
-  border-right: 1px solid var(--border-default);
-  cursor: pointer;
-  transition: all var(--transition-fast);
-  font-family: inherit;
-}
-
-.scope-seg-btn:last-child {
-  border-right: none;
-}
-
-.scope-seg-btn:hover {
-  color: var(--text-primary);
-  background: var(--canvas-subtle);
-}
-
-.scope-seg-btn.active {
-  color: var(--accent-fg);
-  background: var(--accent-muted);
-  font-weight: 600;
-}
-
-/* ── Search ──────────────────────────────────────────────── */
-.search-box {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 0 12px;
-  height: 32px;
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  background: var(--canvas-subtle);
-  max-width: 240px;
-  flex: 1;
-  transition: border-color var(--transition-fast);
-  box-sizing: border-box;
-}
-
-.search-box:focus-within {
-  border-color: var(--accent-fg);
-}
-
-.search-icon {
-  color: var(--text-tertiary);
-  flex-shrink: 0;
-}
-
-.search-input {
-  flex: 1;
-  border: none;
-  background: transparent;
-  color: var(--text-primary);
-  font-size: 0.8125rem;
-  outline: none;
-}
-
-.search-input::placeholder {
-  color: var(--text-tertiary);
-}
-
-/* ── Buttons ─────────────────────────────────────────────── */
-.btn {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 7px 14px;
-  border-radius: var(--radius-md);
-  font-size: 0.8125rem;
-  font-weight: 500;
-  cursor: pointer;
-  white-space: nowrap;
-  transition: all var(--transition-fast);
-  font-family: inherit;
-}
-
-.btn svg {
-  width: 14px;
-  height: 14px;
-}
-
-.btn--primary {
-  background: var(--gradient-accent, var(--accent-emphasis));
-  border: 1px solid transparent;
-  color: var(--text-on-emphasis);
-  font-weight: 600;
-  box-shadow: 0 1px 6px rgba(99, 102, 241, 0.35);
-}
-
-.btn--primary:hover:not(:disabled) {
-  box-shadow: 0 3px 14px rgba(99, 102, 241, 0.45);
-  transform: translateY(-1px);
-}
-
-.btn--primary:active {
-  transform: translateY(0);
-}
-
-.btn--primary:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-  transform: none;
-  box-shadow: none;
-}
-
-.btn--secondary {
-  background: var(--canvas-default, var(--canvas-subtle));
-  border: 1px solid var(--border-default);
-  color: var(--text-secondary);
-}
-
-.btn--secondary:hover {
-  color: var(--text-primary);
-  border-color: var(--accent-fg);
-  background: var(--accent-subtle, var(--canvas-inset));
-}
-
-.btn--ghost {
-  background: transparent;
-  border: 1px solid transparent;
-  color: var(--text-tertiary);
-}
-
-.btn--ghost:hover {
-  color: var(--text-secondary);
-  background: var(--canvas-subtle);
-  border-color: var(--border-default);
-}
-
-.btn--sm {
-  padding: 4px 10px;
-  font-size: 0.75rem;
-}
-
-/* ── Skills Grid ─────────────────────────────────────────── */
-.skills-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-  gap: 12px;
-}
-
-/* ── State Messages ──────────────────────────────────────── */
-.state-message {
-  text-align: center;
-  padding: 40px 20px;
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-}
-
-.state-message--error {
-  color: var(--danger-fg);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-}
-
-/* ── Empty State ─────────────────────────────────────────── */
-.empty-state {
-  text-align: center;
-  padding: 60px 20px;
-}
-
-.empty-state__icon {
-  font-size: 3rem;
-  margin-bottom: 12px;
-}
-
-.empty-state__title {
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin: 0 0 6px;
-}
-
-.empty-state__desc {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-  margin: 0 0 20px;
-}
-
-.empty-state__actions {
-  display: flex;
-  justify-content: center;
-  gap: 8px;
-}
-
-/* ── Modal ───────────────────────────────────────────────── */
-.modal-overlay {
-  position: fixed;
-  inset: 0;
-  z-index: 1000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: rgba(0, 0, 0, 0.6);
-  backdrop-filter: blur(4px);
-}
-
-.modal {
-  width: 440px;
-  max-width: 90vw;
-  background: var(--canvas-default);
-  border-radius: var(--radius-lg);
-  border: 1px solid var(--border-default);
-  box-shadow: var(--shadow-lg);
-}
-
-.modal__header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 16px 20px;
-  border-bottom: 1px solid var(--border-muted);
-}
-
-.modal__title {
-  font-size: 1rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin: 0;
-}
-
-.modal__close {
-  background: none;
-  border: none;
-  color: var(--text-tertiary);
-  cursor: pointer;
-  font-size: 1rem;
-  padding: 4px;
-}
-
-.modal__close:hover {
-  color: var(--text-primary);
-}
-
-.modal__body {
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.modal__label {
-  font-size: 0.75rem;
-  font-weight: 600;
-  color: var(--text-secondary);
-  text-transform: uppercase;
-  letter-spacing: 0.03em;
-}
-
-.modal__input {
-  padding: 8px 12px;
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  background: var(--canvas-subtle);
-  color: var(--text-primary);
-  font-size: 0.8125rem;
-  outline: none;
-  transition: border-color var(--transition-fast);
-}
-
-.modal__input:focus {
-  border-color: var(--accent-fg);
-}
-
-.modal__textarea {
-  padding: 8px 12px;
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  background: var(--canvas-subtle);
-  color: var(--text-primary);
-  font-size: 0.8125rem;
-  outline: none;
-  font-family: inherit;
-  resize: vertical;
-  transition: border-color var(--transition-fast);
-}
-
-.modal__textarea:focus {
-  border-color: var(--accent-fg);
-}
-
-.modal__hint {
-  font-size: 0.75rem;
-  color: var(--text-tertiary);
-  margin: 8px 0 0;
-  font-style: italic;
-}
-
-.modal__validation-hint {
-  font-size: 0.75rem;
-  color: var(--danger-fg);
-  margin: 2px 0 0;
-}
-
-.modal__optional {
-  font-weight: 400;
-  color: var(--text-tertiary);
-  text-transform: none;
-  letter-spacing: 0;
-}
-
-.modal__footer {
-  display: flex;
-  justify-content: flex-end;
-  gap: 8px;
-  padding: 12px 20px;
-  border-top: 1px solid var(--border-muted);
-}
-</style>
