+++ packages/ui/src/components/FileContentViewer.vue
@@ -15,6 +15,7 @@ import { useClipboard } from "../composables/useClipboard";
 import MarkdownContent from "./MarkdownContent.vue";
 import CodeBlock from "./renderers/CodeBlock.vue";
 import SqliteTableView from "./SqliteTableView.vue";
+import "../styles/features/file-content-viewer.css";
 
 const props = defineProps<{
   /** Relative path within the session directory (for display + language detection). */
@@ -311,373 +312,3 @@ function copyTableAsJson() {
     </template>
   </div>
 </template>
-
-<style scoped>
-.fcv {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  min-height: 0;
-  overflow: hidden;
-}
-
-/* ── Empty / placeholder states ──────────────────────────── */
-.fcv__empty,
-.fcv__binary {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  gap: 12px;
-  padding: 32px;
-  text-align: center;
-  color: var(--text-tertiary);
-}
-
-.fcv__empty-icon,
-.fcv__binary-icon {
-  width: 40px;
-  height: 40px;
-  opacity: 0.4;
-}
-
-.fcv__empty-text {
-  font-size: 0.875rem;
-  margin: 0;
-}
-
-.fcv__binary-title {
-  font-size: 0.9375rem;
-  font-weight: 600;
-  color: var(--text-secondary);
-  margin: 0;
-}
-
-.fcv__binary-desc {
-  font-size: 0.8125rem;
-  line-height: 1.5;
-  max-width: 320px;
-  margin: 0;
-}
-
-/* ── Loading ─────────────────────────────────────────────── */
-.fcv__loading {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  height: 100%;
-  font-size: 0.875rem;
-  color: var(--text-tertiary);
-}
-
-.fcv__spinner {
-  width: 18px;
-  height: 18px;
-  border: 2px solid var(--border-default);
-  border-top-color: var(--accent-fg);
-  border-radius: 50%;
-  animation: fcv-spin 0.75s linear infinite;
-}
-
-@keyframes fcv-spin {
-  to { transform: rotate(360deg); }
-}
-
-/* ── Error ───────────────────────────────────────────────── */
-.fcv__error {
-  display: flex;
-  align-items: flex-start;
-  gap: 8px;
-  margin: 16px;
-  padding: 12px 14px;
-  background: var(--danger-muted);
-  border: 1px solid var(--danger-subtle);
-  border-radius: var(--radius-md);
-  color: var(--danger-fg);
-  font-size: 0.8125rem;
-}
-
-.fcv__error-icon {
-  width: 16px;
-  height: 16px;
-  flex-shrink: 0;
-  margin-top: 1px;
-}
-
-.fcv__error-text {
-  line-height: 1.4;
-}
-
-/* ── Copy button ─────────────────────────────────────────── */
-.fcv__copy-btn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 24px;
-  height: 24px;
-  flex-shrink: 0;
-  margin-left: auto;
-  padding: 0;
-  border: none;
-  border-radius: var(--radius-sm);
-  background: transparent;
-  color: var(--text-tertiary);
-  cursor: pointer;
-  transition: color var(--transition-fast), background var(--transition-fast);
-}
-
-.fcv__copy-btn svg {
-  width: 13px;
-  height: 13px;
-}
-
-.fcv__copy-btn:hover {
-  color: var(--text-primary);
-  background: var(--neutral-muted);
-}
-
-.fcv__copy-btn--copied {
-  color: var(--success-fg, #2da44e);
-}
-
-/* ── File header ─────────────────────────────────────────── */
-.fcv__file-header {
-  display: flex;
-  align-items: center;
-  gap: 7px;
-  padding: 8px 12px;
-  height: 36px;
-  box-sizing: border-box;
-  border-bottom: 1px solid var(--border-default);
-  background: var(--canvas-subtle);
-  flex-shrink: 0;
-}
-
-.fcv__file-icon {
-  width: 13px;
-  height: 13px;
-  color: var(--text-tertiary);
-  opacity: 0.7;
-  flex-shrink: 0;
-}
-
-.fcv__file-name {
-  font-size: 0.75rem;
-  font-family: var(--font-mono);
-  color: var(--text-secondary);
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-/* ── Content area ────────────────────────────────────────── */
-.fcv__content {
-  flex: 1;
-  min-height: 0;
-  overflow-y: auto;
-  padding: 0;
-}
-
-/* When a CodeBlock owns the scroll (fill-height mode), this container must
-   not scroll itself — it just passes the bounded height down to the child. */
-.fcv__content--fill {
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-}
-
-.fcv__content :deep(.markdown-content) {
-  padding: 16px 20px;
-}
-
-/* ── SQLite database viewer ──────────────────────────────── */
-.fcv__db-header {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 8px 12px;
-  height: 36px;
-  box-sizing: border-box;
-  border-bottom: 1px solid var(--border-default);
-  background: var(--canvas-subtle);
-  flex-shrink: 0;
-}
-
-.fcv__db-icon {
-  width: 14px;
-  height: 14px;
-  color: var(--accent-fg);
-  flex-shrink: 0;
-}
-
-.fcv__db-name {
-  font-size: 0.75rem;
-  font-family: var(--font-mono);
-  color: var(--text-secondary);
-  flex: 1;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.fcv__db-meta {
-  font-size: 0.6875rem;
-  color: var(--text-tertiary);
-  flex-shrink: 0;
-}
-
-.fcv__db-tabs {
-  display: flex;
-  align-items: stretch;
-  gap: 2px;
-  padding: 4px 8px 0;
-  border-bottom: 1px solid var(--border-default);
-  background: var(--canvas-default);
-  flex-shrink: 0;
-  overflow-x: auto;
-  scrollbar-width: thin;
-}
-
-.fcv__db-tabs-spacer {
-  flex: 1 1 auto;
-  min-width: 8px;
-}
-
-.fcv__db-view-toggle {
-  display: inline-flex;
-  gap: 2px;
-  padding: 2px;
-  margin: 2px 0 4px;
-  background: var(--surface-secondary, var(--canvas-subtle));
-  border: 1px solid var(--border-default);
-  border-radius: 999px;
-  align-self: center;
-  flex-shrink: 0;
-}
-
-.fcv__db-view-btn {
-  padding: 3px 10px;
-  border: none;
-  border-radius: 999px;
-  background: transparent;
-  color: var(--text-secondary);
-  font-size: 0.6875rem;
-  font-weight: 500;
-  cursor: pointer;
-  white-space: nowrap;
-  transition: background var(--transition-fast), color var(--transition-fast);
-}
-
-.fcv__db-view-btn:hover:not(.fcv__db-view-btn--active) {
-  color: var(--text-primary);
-  background: var(--canvas-inset, var(--canvas-subtle));
-}
-
-.fcv__db-view-btn--active {
-  background: var(--accent-emphasis, var(--accent-fg));
-  color: var(--text-on-emphasis, #fff);
-}
-
-.fcv__db-tab {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  padding: 5px 10px;
-  border: none;
-  border-bottom: 2px solid transparent;
-  background: none;
-  font-size: 0.75rem;
-  color: var(--text-secondary);
-  cursor: pointer;
-  white-space: nowrap;
-  border-radius: var(--radius-sm) var(--radius-sm) 0 0;
-  transition: color var(--transition-fast), background var(--transition-fast);
-}
-
-.fcv__db-tab:hover {
-  color: var(--text-primary);
-  background: var(--canvas-subtle);
-}
-
-.fcv__db-tab--active {
-  color: var(--accent-fg);
-  border-bottom-color: var(--accent-fg);
-  font-weight: 500;
-}
-
-.fcv__db-tab-count {
-  font-size: 0.6875rem;
-  padding: 1px 5px;
-  border-radius: 99px;
-  background: var(--neutral-muted);
-  color: var(--text-secondary);
-}
-
-.fcv__db-content {
-  flex: 1;
-  overflow: auto;
-  min-height: 0;
-  display: flex;
-  flex-direction: column;
-}
-
-.fcv__db-empty {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  font-size: 0.8125rem;
-  color: var(--text-tertiary);
-}
-
-.fcv__db-table-wrap {
-  overflow: auto;
-  padding: 0;
-}
-
-.fcv__db-table {
-  width: 100%;
-  border-collapse: collapse;
-  font-size: 0.75rem;
-  font-family: var(--font-mono);
-}
-
-.fcv__db-th {
-  position: sticky;
-  top: 0;
-  padding: 7px 10px;
-  text-align: left;
-  font-weight: 600;
-  background: var(--canvas-subtle);
-  border-bottom: 1px solid var(--border-default);
-  color: var(--text-secondary);
-  white-space: nowrap;
-  z-index: 1;
-}
-
-.fcv__db-tr:nth-child(even) {
-  background: var(--canvas-subtle);
-}
-
-.fcv__db-tr:hover {
-  background: var(--neutral-muted);
-}
-
-.fcv__db-td {
-  padding: 5px 10px;
-  border-bottom: 1px solid var(--border-muted);
-  max-width: 300px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  vertical-align: top;
-  color: var(--text-primary);
-}
-
-.fcv__db-null {
-  color: var(--text-tertiary);
-  font-style: italic;
-}
-</style>
