+++ scripts/check-no-backdrop-filter.mjs
@@ -31,10 +31,10 @@ const ALLOW_FILES = new Set([
   "apps/desktop/src/styles/features/sdk-steering.css",
   "apps/desktop/src/styles/features/session-launcher.css",
   "apps/desktop/src/styles/features/skill-import-wizard.css",
+  "apps/desktop/src/styles/features/skills-manager.css",
   "apps/desktop/src/styles/features/worktree-manager.css",
   "apps/desktop/src/views/SessionListView.vue",
   "apps/desktop/src/views/tabs/ConversationTab.vue",
-  "apps/desktop/src/views/skills/SkillsManagerView.vue",
   "apps/desktop/src/components/WhatsNewModal.vue",
   "apps/desktop/src/components/UpdateInstructionsModal.vue",
   "apps/desktop/src/components/session/SessionDetailPanel.vue",
